+++ plugins/UsersManager/UsersManager.php
@@ -25,6 +25,7 @@ use Piwik\SettingsPiwik;
 class UsersManager extends \Piwik\Plugin
 {
     const PASSWORD_MIN_LENGTH = 6;
+    const PASSWORD_MAX_LENGTH = 200;
 
     /**
      * @see \Piwik\Plugin::registerEvents
@@ -185,6 +186,9 @@ class UsersManager extends \Piwik\Plugin
 
         if (!self::isValidPasswordString($password)) {
             throw new Exception(Piwik::translate('UsersManager_ExceptionInvalidPassword', array(self::PASSWORD_MIN_LENGTH)));
+        }
+        if (Common::mb_strlen($password) > self::PASSWORD_MAX_LENGTH) {
+            throw new Exception(Piwik::translate('UsersManager_ExceptionInvalidPasswordTooLong', array(self::PASSWORD_MAX_LENGTH)));
         }
     }