+++ core/Metrics.php
@@ -242,6 +242,21 @@ class Metrics
             '_time_'  => 's'
         );
 
+        $unit = null;
+
+        /**
+         * Use this event to define units for custom metrics used in evolution graphs and row evolution only.
+         *
+         * @param string $unit should hold the unit (e.g. %, €, s or empty string)
+         * @param string $column name of the column to determine
+         * @param string $idSite id of the current site
+         */
+        Piwik::postEvent('Metrics.getEvolutionUnit', [&$unit, $column, $idSite]);
+
+        if (!empty($unit)) {
+            return $unit;
+        }
+
         foreach ($nameToUnit as $pattern => $type) {
             if (strpos($column, $pattern) !== false) {
                 return $type;